Update principal_filter_analysis.ipynb
Greetings,
Please find below a list of minor changes recommended to the documentation.

Line 56: Replace '1a' with ' 1.a' to follow the notebook's prevalent format

Line 113:
i) Replace 'Similarly' with 'Similar'
ii) The sentence, "This is the set of layers that to analyze using PFA", lacks clarity. This could be rephrased as "This is the set of layers that will be used to analyze with PFA". (Please note that my interpretation of the program could be wrong.)
iii) Replace '1.f' with '1.e'

Line 149: Add 'that'

Line 193: '1.f' has been used when there is no prior '1.e'. Hence, replace '1.f' with '1.e'

Line 291: Add 'be' in between 'also' and 'queried'

Thank you